core/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
Brotli==1.1.0
cachetools==4.2.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.7.0
django-cors-headers==4.6.0
django-googledrive-storage==1.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.2.0
google-api-core==2.10.2
google-api-python-client==2.172.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
ipython==8.32.0
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.50
protobuf==4.25.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rpds-py==0.22.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.13
whitenoise==6.9.0
